+++ riscv/CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Use CSR helper macros to define `mcause` register
 - Use CSR helper macros to define `medeleg` register
 - Use CSR helper macros to define `mideleg` register
+- Use CSR helper macros to define `mcounteren` register
 
 ## [v0.12.1] - 2024-10-20